+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Renamed method at experimental API reader.PopBatchTx to reader.PopMessagesBatchTx
+
 ## v3.80.5
 * Fixed connections pool leak on failed `ydb.Open` call